requirements.txt
numpy<2.0
Flask==2.2.5
Pillow==9.2.0
prettytable==2.0.0
requests==2.25.1
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
tqdm==4.49.0
wcwidth==0.2.8
Werkzeug==2.2.3
opencv-python
